SplashScreen was done
The SplashScreen was done by using a theme as an approach. This approach basically
does the SplashScreen waits for the next screen to be entirely prepared before
finish itself.

The best point in this approach is that we don't need to be concerned about
the time the OS will need to do its tasks. The OS always will wait the second
view signalize it is prepared to be launched to end the SplashScreen.