Rest config initiated
In this commit, some files were added to initiate the rest consumption of the API.
We will use the MovieDB API to get some movie info from there. To accomplish that task,
we will use Retrofit 2 and OkHTTP 3 to make the communication between the server and us.

Minor adjustments were done too and they are described as follow:

- The .keystore file for signing the app for release was added;
- 2 interfaces were designed to receive possible anonymous and authenticated requests;
- Proguard was modified to accept Retrofit, Picasso and Okio's files on release;
- A suffix was added in front of the applicationId for development.